ajustaes varios

+++ ejbModule/py/edu/ucsa/ejb/dao/IEquipoDao.java
@@ -2,7 +2,7 @@ package py.edu.ucsa.ejb.dao;
 
 import py.edu.ucsa.ejb.entities.Equipo;
 
-public interface IEquipoDao extends IGenericDao<Long, Equipo> {
+public interface IEquipoDao extends IGenericDao<Integer, Equipo> {
 
 	public Iterable<Equipo> findAllByNombre(String nombre, boolean isNull);
 }
