+++ ejbModule/py/edu/ucsa/ejb/dao/IJugadorDao.java
@@ -2,7 +2,7 @@ package py.edu.ucsa.ejb.dao;
 
 import py.edu.ucsa.ejb.entities.Jugador;
 
-public interface IJugadorDao extends IGenericDao<Long, Jugador> {
+public interface IJugadorDao extends IGenericDao<Integer, Jugador> {
 
 	public Iterable<Jugador> findByNombre(String nombre, boolean isNull);
 }
