+++ ejbModule/py/edu/ucsa/ejb/dao/ITorneoDao.java
@@ -2,6 +2,6 @@ package py.edu.ucsa.ejb.dao;
 
 import py.edu.ucsa.ejb.entities.Torneo;
 
-public interface ITorneoDao extends IGenericDao<Long, Torneo> {
-	public Iterable<Torneo> findByAno(int ano);
+public interface ITorneoDao extends IGenericDao<Integer, Torneo> {
+	public Iterable<Torneo> findByAno(Integer ano);
 }
